Get performance data only for interfaces matching the regex
I think it was intended this way but the code was misordered, so it did not work as expected (instead all interfaces were polled which is not really efficient especially witch many Interfaces). I also implemented a bulkget for the if-names, so we can match them with the regex too. I could only test it with a cisco switch and I do not know whether it works for Bintec.