Check for availability of DES privProtocol during compile time
Some distributors (Red Hat) choose to remove DES support from net-snmp
completely (for security reasons probably).
This made the compilations of this program fail, although DES is
optional here (one option of three).

This change tries to check for the availability of DES in libsnmp
and includes it if possible.