Change toolchain
This commit puts the main function in a separate check_interfaces.c
to make it more obvious what happens.
The code is now (badly) separated from snmp_bulkget.c and needs further
logical separation

configure.ac is enhanced with this commit and the Makefile generation
is now done from a Makefile.am, so the Makefile.in in this repository
is theoretically obsolete.

Some other (for now empty) files in here are just generated to satisfy
the needs of the autoconf toolchain