Try (harder?) to detect the real number of check_interfaces
It seems like detecting the number of interfaces a device has is harder
than one might think.
Apparently those things lie HARD about that.
If you then try to allocate memory for the result based on that,
but try to write MORE results there, you get a segfault sooner or later.

This patch iterates through the answer of the device and just counts
them, so this is hopefully the safe way to do this.